Fix loading cached call
The hash calculated in the Call object is different from the hash that
is used to fetch a cached call. The Call object includes the API version
in the hash which is not the case for the hash to fetch a cached call.
Now the hash from the Call object is used to fetch a cached call.